move getAbsoluteUrl to magda-typescipt-common
